rewrite allocator

+++ src/c/Memory.h
@@ -26,7 +26,7 @@ void Memory_delete(unsigned int size);
 void Memory_size(unsigned int i);
 
 
-void* malloc(long num_bytes);
+void* malloc(u32 num_bytes);
 void free(void* ptr);
 void memcpy();
 void memset();
